Finalise Service

+++ VCharge.Models/MeterReading.cs
@@ -5,6 +5,6 @@ namespace VCharge.Models
     public class MeterReading
     {
         public DateTime Date { get; set; }
-        public double CumulativeConsumption { get; set; }
+        public decimal CumulativeConsumption { get; set; }
     }
 }
