+++ VCharge.Models/MonthlySummary.cs
@@ -6,7 +6,7 @@ namespace VCharge.Models
     public class MonthlySummary
     {
         public DateTime DateMonthStart { get; set; }
-        public double KwhUsageAtMonthStart { get; set; }
-        public double KwhUsageAtMonthEnd { get; set; }
+        public decimal KwhUsageAtMonthStart { get; set; }
+        public decimal KwhUsageAtMonthEnd { get; set; }
     }
 }
