+++ VCharge.Services/IMeterReadingAggregationService.cs
@@ -7,6 +7,6 @@ namespace VCharge.Services
     public interface IMeterReadingAggregationService
     {
         IEnumerable<MonthlySummary> GetMonthlyData(IEnumerable<MeterReading> dayReadings);
-        double GetUsageBetweenDates(IEnumerable<MeterReading> dayReadings, DateTime startDate, DateTime endDate);
+        decimal GetUsageBetweenDates(IEnumerable<MeterReading> dayReadings, DateTime startDate, DateTime endDate);
     }
 }
